bug(sys/storage): try to decrease braking time
BUG:
When passing powercells for shooting, sometime the next powercell will
also pass after shooting the original, but at lower speed

CAUSE:
Setting 0 takes a bit too long to completley stop the motor, and so the
next powercell is passed while the storaege system is slowing down.

FIX:
Setting the netral mode of the storage controller to break should try to
slow it down more actively with a counter acceleration, and thus
decrease the stopping time.

NOTES:
This is currently untested.